backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
psycopg2-binary==2.9.7
apscheduler==3.10.4
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
email-validator==2.0.0 